backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
httpx==0.26.0
cryptography==41.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyMuPDF==1.26.0
reportlab==4.4.0
sqladmin==0.23.0
email-validator==2.1.0
